requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
Tempita==0.5.2
Werkzeug==0.9.6
decorator==3.4.0
itsdangerous==0.24
pbr==0.10.0
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.12
wsgiref==0.1.2
